+++ components/clarinet-cli/src/runner/mod.rs
@@ -48,6 +48,7 @@ impl DeploymentCache {
             &deployment,
             Some(&artifacts.asts),
             true,
+            None,
         );
 
         let mut contracts_artifacts = HashMap::new();
@@ -142,7 +143,7 @@ pub fn block_on<F, R>(future: F) -> R
 where
     F: std::future::Future<Output = R>,
 {
-    let rt = crate::utils::create_basic_runtime();
+    let rt = hiro_system_kit::create_basic_runtime();
     rt.block_on(future)
 }
 pub struct SessionArtifacts {